+ the developer version for gh-pages
This adresses the issue #4 gh-pages. Could someone please review the changes and rename the branch if it is okay.
I tried to do as much as I can on a static site to prevent search engines from crawling the site. Now you will need JS to look at the site and its content. (or you just manualy alter the css but okay)
Additionally I added a reveal box with a message stating that this is just a developer site.