+ added a cookie to prevent annoying reveal
I just noticed that with every reload the reveal gets triggered. I did not thought of this and this is not okay. So I've written a little cookie setting into the JS. The cookie now expires 5 days from clicking the approval. We can change this value if neccessary.
This also belongs to issue #4