* fixing timetable overflow issue
I think that should be it. The date was too long and did not break so I just added a short version (15 instead of 2015) of the date in the events.yaml and now it works. It still overflows the main content area but as far as I can see it doesn't conflict with the sidebar so this should be the final version. relatiert #7